hornbeam_callbacks (hornbeam v1.0.0)

Erlang callback registry for Python apps.

This module manages callbacks that Python apps can invoke. It wraps the erlang-python function registration with additional features like validation and logging.

Functions

call(Name, Args)

-spec call(Name :: atom() | binary(), Args :: list()) -> {ok, term()} | {error, term()}.

Call a registered callback synchronously.

cast(Name, Args)

-spec cast(Name :: atom() | binary(), Args :: list()) -> ok.

Call a registered callback asynchronously (fire and forget).

code_change(OldVsn, State, Extra)

handle_call(Request, From, State)

handle_cast(Request, State)

handle_info(Info, State)

init(_)

list()

-spec list() -> [atom() | binary()].

List all registered callbacks.

register(Name, Fun)

-spec register(Name :: atom() | binary(), Fun :: fun((list()) -> term())) -> ok.

Register a callback function. The function should accept a list of arguments.

register(Name, Module, Function)

-spec register(Name :: atom() | binary(), Module :: atom(), Function :: atom()) -> ok.

Register a module:function as a callback.

start_link()

Start the callback manager.

terminate(Reason, State)

unregister(Name)

-spec unregister(Name :: atom() | binary()) -> ok.

Unregister a callback.